backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
SQLAlchemy==2.0.43
PyMySQL==1.1.1
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
